// src/compositions/cricket/utils/broadcastProRounded/playerRanking/types.ts
export interface BroadcastProRoundedTripleStat {
  label1: string;
  value1: string;
  label2: string;
  value2: string;
  label3: string;
  value3: string;
}

export type BroadcastProRoundedPlayerRankingLayout =
  | "featured"
  | "gridTop5"
  | "gridPerformances";

export type BroadcastProRoundedPlayerRankBadgePlacement = "left" | "right";